requirements.txt
llama-index
dnspython-2.6.1 
pymongo-4.7.0